pyproject.toml
[project]
name = "noticias-bot"
version = "0.1.0"
description = "Bot de Telegram con IA para resumir noticias en tiempo real"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "httpx>=0.28.0",
    "instructor>=1.14.5",
    "pydantic>=2.12.5",
    "openai>=2.30.0",
    "python-telegram-bot>=22.7",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.15.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/noticias_bot"]


[tool.ruff]
# Configuración moderna de Ruff para 2026
line-length = 88
target-version = "py313"
lint.select = ["E", "F", "I", "B"] # Errores, Fallos, Imports, Bugs comunes

[tool.pyright]
# Le indicamos a Pyright dónde buscar nuestro código
include = ["src"]
extraPaths = ["src"]
typeCheckingMode = "strict"
pythonVersion = "3.13"
